move the input handling into its own file in the module,move InitRenderer from gl.go to hypervisor,change rune to uint32
