+++ hypervisor/gl.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"github.com/corpusc/viscript/app"
 	"github.com/corpusc/viscript/gfx"
-	"github.com/corpusc/viscript/script"
 	"github.com/go-gl/gl/v2.1/gl"
 	"image"
 	"image/draw"
@@ -18,37 +17,6 @@ var (
 	rotationX float32
 	rotationY float32
 )
-
-func InitRenderer() {
-	fmt.Println("InitRenderer()")
-
-	gl.Enable(gl.DEPTH_TEST)
-	gl.Enable(gl.LIGHTING)
-	//gl.Enable(gl.ALPHA_TEST)
-
-	gl.ClearColor(0.5, 0.5, 0.5, 0.0)
-	gl.ClearDepth(1)
-	gl.DepthFunc(gl.LEQUAL)
-
-	ambient := []float32{0.5, 0.5, 0.5, 1}
-	diffuse := []float32{1, 1, 1, 1}
-	lightPosition := []float32{-5, 5, 10, 0}
-	gl.Lightfv(gl.LIGHT0, gl.AMBIENT, &ambient[0])
-	gl.Lightfv(gl.LIGHT0, gl.DIFFUSE, &diffuse[0])
-	gl.Lightfv(gl.LIGHT0, gl.POSITION, &lightPosition[0])
-	gl.Enable(gl.LIGHT0)
-
-	gl.MatrixMode(gl.PROJECTION)
-	gl.LoadIdentity()
-	setFrustum(gfx.InitFrustum)
-	//gl.Frustum(-1, 1, -1, 1, 1.0, 10.0)
-	//gl.Frustum(left, right, bottom, top, zNear, zFar)
-	gl.MatrixMode(gl.MODELVIEW)
-	gl.LoadIdentity()
-
-	// future FIXME: finished app would not have a demo program loaded on startup?
-	script.Process(false)
-}
 
 func setFrustum(r *app.Rectangle) {
 	gl.Frustum(
